Improve WASM generated size
The fix is passing `-C link-args=-s` as flag to the compiler.
The current attempt was setting `WASM_BUILD_RUSTFLAGS` but this is a
custom environment var of parity that has no effect whatsoever in
a default environment.

Now we call directly `cargo rustc` in the Makefile, with all parameters
needed.
I also add few optimization in the contract's Cargo.toml (`lto` and
`codegen-units`) to skimmer other few KBs.

For `transfer` contract we went from ~680KB to 16KB (`-C link-args=-s`)
to 14KB (`lto` / `codegen-units`).

We still need to strip down the panicking infrastructure.

Resolves: #34, #35
See also: #36