Implement bytecode instrumentation at deploy time
As stated in #115 and #116 we need a way to instrumentate the WASM
bytecode of the contracts deployed to be able to inject gas_metering
tools as well as removal of floating point operations among other
checks.

This PR adds the tooling and instrumentation linked to the
`NetworksState::deploy()` fn adding the following checks and
instrumentation for the bytecode:
- No floating point ops are used in the contract.
- Gas metering tools are injected correctly and work.
- Stack height metering tools so that we never pass the maximum allowed.
- Table size limits used in the wasm Modules are not bigger than what is
allowed.
- Validity check of all of the previous instrumentation.

Resolves: #115, #116.