Remove duplicated floating point filtering logic
Without noticing, it was introduced twice the logic to filter floating
point operators inside of WASM. One inside of `inject_gas_metering` and
the other in `ensure_no_floating_types`.

The second one has been removed since the same code is implicit in the
code that the first function executes inside on it's original lib
`wasm_utils`.