Remove ContractInstrumenter struct
The existance of this struct was not needed and seemed too much to the
classical Java way of doing things where you create a Class to have any
extra method definition.

Also, there was no real need to instanciate the struct at all since we
just wanted the functions inside of it to be executed, nothing else.