Add `gas_consumed` host function logic
In dusk-network/dusk-abi#25 we introduced the
interface required so that the contracts can obtain the current
gas_consumed until a certain point of the execution.

This is needed specifically by the transfer contract as stated in #174
so that the contract can know how much Dusk the `Fee` that will be
charged to the transactor needs to contain.

To do so, this introduces the logic of the `gas_consumed` inter contract
call as well as a new contract to test that the function works as
expected.

Resolves: #174