Change the way `NetworkState` gets its configurations
`NetworkState` used to take the module configurations from a
`ModuleConfig` struct, which in turn read the configurations from a
`Schedule` struct, which contained our hard-coded configurations.
To make this process more transparent and adjustable, we deprecated
`Schedule` and now handle all configurations in the `config`
(previously `module_config`) module.
With this setup, we can read configurations from file or generate
a set of sensible default configurations.

Also change `InstrumentalizationError` to `InstrumentationError`.

Resolves: #221