Remove rust toolchain overrides from CI
The test was failing because a more recent nightly was used in the CI,
instead of using the version specified in `rust-toolchain`.
Locally everything was working properly because the overrides was on the
CI config file.

Without the override both CI and local execution are using the same
version specified in `rust-toolchain`.

Resolves: #229